Enable persisting battery data in db

+++ common/src/main/java/xyz/ruhshan/pl/common/entity/Battery.java
@@ -14,5 +14,5 @@ import lombok.NoArgsConstructor;
 public class Battery extends BaseEntity{
     private String name;
     private Integer postCode;
-    private Integer capacity;
+    private Double capacity;
 }
